requirements.txt
flask==3.0.2
python-dotenv==1.0.1
bcrypt==4.1.2
pandas==2.2.1
yfinance==0.2.37
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.1
plotly==5.19.0
dash==2.15.0
python-jose==3.3.0 